:building_construction: 로그인 인증 관련 함수들 분리하여 클래스 생성
